stars being shown on existing reviews

+++ reviews/models.py
@@ -10,7 +10,7 @@ class Review(models.Model):
     username = models.CharField(max_length=200)
     review_title = models.CharField(max_length=200)
     review_content = models.TextField()
-    rating = models.DecimalField(max_digits=2, decimal_places=1, default=0)
+    rating = models.IntegerField(default=0)
     review_date = models.DateTimeField(default=datetime.now, blank=True)
     user_id = models.IntegerField(blank=True)
     user_image_url = models.CharField(max_length=300, blank=True)
